# JIT: Optimize creation of binaries