--- a/erts/emulator/beam/erl_bits.h
+++ b/erts/emulator/beam/erl_bits.h
@@ -51,23 +51,20 @@ typedef struct erl_bin_match_buffer {
 
 struct erl_bits_state {
     /*
-     * Used for building binaries.
+     * Temporary buffer sometimes used by erts_new_bs_put_integer().
      */
     byte *byte_buf_;
     int byte_buf_len_;
+
     /*
-     * Used for building binaries using the new instruction set.
+     * Pointer to the beginning of the current binary.
      */
-    byte* erts_current_bin_;	/* Pointer to beginning of current binary. */
+    byte* erts_current_bin_;
+
     /*
-     * Offset in bits into the current binary (new instruction set) or
-     * buffer (old instruction set).
+     * Offset in bits into the current binary.
      */
     Uint erts_bin_offset_;
-    /*
-     * Whether the current binary is writable.
-     */
-     unsigned erts_writable_bin_;
 };
 
 typedef struct erl_bin_match_struct{
@@ -117,7 +114,6 @@ typedef struct erl_bin_match_struct{
 
 #define erts_bin_offset		(ErlBitsState.erts_bin_offset_)
 #define erts_current_bin	(ErlBitsState.erts_current_bin_)
-#define erts_writable_bin       (ErlBitsState.erts_writable_bin_)
 
 #define copy_binary_to_buffer(DstBuffer, DstBufOffset, SrcBuffer, SrcBufferOffset, NumBits) \
   do {											    \