… erts/emulator/beam/jit/arm/instr_bs.cpp
… erts/emulator/beam/jit/x86/instr_bs.cpp